+++ packages/iov-ethereum/src/encoding.ts
@@ -40,6 +40,13 @@ export function eip155V(chain: Eip155ChainId, recoveryParam: number): number {
 }
 
 export function getRecoveryParam(chain: Eip155ChainId, v: number): number {
+  // After the implementation of EIP-155, clients are still free to use the old
+  // way of calculating v does not protect their users against replay attacks.
+  // https://ethereum.stackexchange.com/a/23955
+  if (v === 27 || v === 28) {
+    return v - 27;
+  }
+
   if (chain.forkState === BlknumForkState.Forked && chain.chainId > 0) {
     // chain ID available
     const recoveryParam = new Int53(v - chain.chainId * 2 - 35);
@@ -51,5 +58,6 @@ export function getRecoveryParam(chain: Eip155ChainId, v: number): number {
     }
     return recoveryParam.toNumber();
   }
+
   throw new Error("transaction not supported before eip155 implementation");
 }