+++ packages/iov-crypto/types/secp256k1.d.ts
@@ -8,4 +8,5 @@ export declare class Secp256k1 {
     static makeKeypair(privkey: Uint8Array): Promise<Secp256k1Keypair>;
     static createSignature(messageHash: Uint8Array, privkey: Uint8Array): Promise<ExtendedSecp256k1Signature>;
     static verifySignature(signature: Secp256k1Signature, messageHash: Uint8Array, pubkey: Uint8Array): Promise<boolean>;
+    static recoverPubkey(signature: ExtendedSecp256k1Signature, messageHash: Uint8Array): Uint8Array;
 }