+++ packages/iov-ethereum/src/encoding.ts
@@ -1,5 +1,7 @@
 import * as rlp from "rlp";
 
+import { Int53 } from "@iov/encoding";
+
 /**
  * Encode as RLP (Recursive Length Prefix)
  */
@@ -38,9 +40,24 @@ export function eip155V(chain: Eip155ChainId, recoveryParam: number): number {
 }
 
 export function getRecoveryParam(chain: Eip155ChainId, v: number): number {
+  // After the implementation of EIP-155, clients are still free to use the old
+  // way of calculating v does not protect their users against replay attacks.
+  // https://ethereum.stackexchange.com/a/23955
+  if (v === 27 || v === 28) {
+    return v - 27;
+  }
+
   if (chain.forkState === BlknumForkState.Forked && chain.chainId > 0) {
     // chain ID available
-    return v - chain.chainId * 2 - 35;
+    const recoveryParam = new Int53(v - chain.chainId * 2 - 35);
+    if (recoveryParam.toNumber() < 0 || recoveryParam.toNumber() > 3) {
+      throw new Error(
+        `Calculated recovery parameter must be one of 0, 1, 2, 3 but is ${recoveryParam}. ` +
+          `Got v: ${v} and chain ID: ${chain.chainId}`,
+      );
+    }
+    return recoveryParam.toNumber();
   }
+
   throw new Error("transaction not supported before eip155 implementation");
 }